Implement Applicability
Actionability
-------------
Add the ability to use a regular expression to determine if the project is
applicable if any file in the project matches the regular expression.
Add the ability to combine multiple applicability conditions.

Project
-------
Add a MonadProject abstraction that manages listing files in the project.
Create a ProjectContext that stores the project folder we get from Lift and
implement MonadProject on a ReaderT instance with a ProjectContext.
Add an allowed-orphans implementation of MonadProject via HMock in the test package.