Wire up applicability and fix bugs
The application now requires a configuration that includes an applicability
condition. The example is updated to search for rust cargo projects.

Export the regex library's types that we use for applicability for convenience.

Remove an invalid "MonadProject IO" and fix the test that accidentally depended
on it.

Remove debugging print statements.